requirements.txt
pydantic
python-multipart
pyflakes
fastapi>=0.112
uvicorn>=0.30
python-dotenv>=1.0
pyyaml>=6.0
requests>=2.32
Jinja2>=3.1
azure-identity>=1.17
# Prefer GA/stable where available; pin to known-working newer versions
azure-ai-projects>=1.0.0
azure-ai-agents>=1.1.0

# Tracing/telemetry (optional but recommended)
opentelemetry-sdk>=1.25
azure-monitor-opentelemetry>=1.6.3
azure-core-tracing-opentelemetry>=1.0.0b11
# Optional local viewer exporter
opentelemetry-exporter-otlp>=1.25